freetds/macros.h: Tweak TDS_INT2PTR for modern Clang.
Formally cast the argument to intptr_t before proceeding, to ensure
proper sign extension under newer Clang versions (under which the
rowset test otherwise fails).  #include "tds_sysdep_public.h"
accordingly; drop now-redundant inclusions thereof from anything that
includes macros.h (notably tds.h).

+++ include/freetds/macros.h
@@ -29,6 +29,8 @@
 #include <stddef.h>
 #endif /* HAVE_STDDEF_H */
 
+#include "tds_sysdep_public.h"
+
 #define TDS_ZERO_FREE(x) do {free((x)); (x) = NULL;} while(0)
 #define TDS_VECTOR_SIZE(x) (sizeof(x)/sizeof(x[0]))
 
@@ -70,7 +72,7 @@
 #define TDS_WUR
 #endif
 
-#define TDS_INT2PTR(i) ((void*)(((char*)0)+(i)))
+#define TDS_INT2PTR(i) ((void*)(((char*)0)+((intptr_t)(i))))
 #define TDS_PTR2INT(p) ((int)(((char*)(p))-((char*)0)))
 
 #endif
